requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic[email]==2.6.0
PyJWT==2.8.0
python-multipart==0.0.6
